First Dataset for the Chatbot Archive added
Lots more work to do to this dataset, but for now this will help me work through the code to make sure everything works. This dataset is complete in the notion that it has all chatbots included, but incomplete in the notion that it does not yet have all of the information for all of the chatbots yet.